# api/swagger.yml
...
consumes:
- application/json
produces:
- application/json
swagger: "2.0"
info:
  description: API documentation
  title: github.com/hzbay/chain-bridge
  version: 0.1.0
paths:
...
  /-/chains/{chain_id}:
    get:
      security:
      - Management: []
      description: Retrieves detailed configuration for a specific blockchain network
      tags:
      - cpop mgmt
      summary: Get chain configuration by ID
      operationId: getChainConfig
      parameters:
      - type: integer
        format: int64
        description: The blockchain network ID
        name: chain_id
        in: path
        required: true
      responses:
        "200":
          description: Successfully retrieved chain configuration
          schema:
            $ref: '#/definitions/ChainResponse'
        "400":
          description: Bad request error
          schema:
            $ref: '#/definitions/apiError'
          examples:
            application/json:
              message: Bad request
        "404":
          description: Not found error
          schema:
            $ref: '#/definitions/apiError'
          examples:
            application/json:
              message: Not found
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apiError'
          examples:
            application/json:
              message: Internal server error
    put:
      security:
      - Management: []
      description: Updates the configuration for a specific blockchain network
      tags:
      - cpop mgmt
      summary: Update chain configuration
      operationId: updateChain
      parameters:
      - type: integer
        format: int64
        description: The blockchain network ID
        name: chain_id
        in: path
        required: true
      - description: Chain update request
        name: request
        in: body
        required: true
        schema:
          $ref: '#/definitions/updateChainRequest'
      responses:
        "200":
          description: Chain updated successfully
          schema:
            $ref: '#/definitions/chainUpdateResponse'
        "400":
          description: Bad request error
          schema:
            $ref: '#/definitions/apiError'
          examples:
            application/json:
              message: Bad request
        "404":
          description: Not found error
          schema:
            $ref: '#/definitions/apiError'
          examples:
            application/json:
              message: Not found
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apiError'
          examples:
            application/json:
              message: Internal server error
    delete:
      security:
      - Management: []
      description: Removes a blockchain network from the supported chains list
      tags:
      - cpop mgmt
      summary: Delete chain configuration
      operationId: deleteChain
      parameters:
      - type: integer
        format: int64
        description: The blockchain network ID
        name: chain_id
        in: path
        required: true
      responses:
        "200":
          description: Chain deleted successfully
          schema:
            $ref: '#/definitions/chainDeleteResponse'
        "400":
          description: Bad request error
          schema:
            $ref: '#/definitions/apiError'
          examples:
            application/json:
              message: Bad request
        "404":
          description: Not found error
          schema:
            $ref: '#/definitions/apiError'
          examples:
            application/json:
              message: Not found
        "409":
          description: Chain is in use and cannot be deleted
          schema:
            $ref: '#/definitions/apiError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apiError'
          examples:
            application/json:
              message: Internal server error
...
  /-/chains/{chain_id}/toggle-status:
    put:
      security:
      - Management: []
      description: Enables or disables a specific blockchain network
      tags:
      - cpop mgmt
      summary: Toggle chain enabled status
      operationId: toggleChainStatus
      parameters:
      - type: integer
        format: int64
        description: The blockchain network ID
        name: chain_id
        in: path
        required: true
      - type: boolean
        description: Whether to enable or disable the chain
        name: enabled
        in: query
        required: true
      responses:
        "200":
          description: Chain status updated successfully
          schema:
            $ref: '#/definitions/chainToggleStatusResponse'
        "400":
          description: Bad request error
          schema:
            $ref: '#/definitions/apiError'
          examples:
            application/json:
              message: Bad request
        "404":
          description: Not found error
          schema:
            $ref: '#/definitions/apiError'
          examples:
            application/json:
              message: Not found
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apiError'
          examples:
            application/json:
              message: Internal server error
...
  /-/tokens:
    get:
      security:
      - Management: []
      description: Retrieves the list of all supported tokens across all chains
      tags:
      - cpop mgmt
      summary: Get all supported tokens
      operationId: getTokens
      parameters:
      - type: integer
        format: int64
        description: Filter tokens by chain ID
        name: chain_id
        in: query
      - type: boolean
        default: false
        description: Only return enabled tokens
        name: enabled_only
        in: query
      - type: boolean
        default: false
        description: Only return tokens that support batch operations
        name: batch_support_only
        in: query
      responses:
        "200":
          description: Successfully retrieved tokens
          schema:
            $ref: '#/definitions/tokensResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apiError'
          examples:
            application/json:
              message: Internal server error
    post:
      security:
      - Management: []
      description: Adds a new token to the supported tokens list
      tags:
      - cpop mgmt
      summary: Create new supported token
      operationId: createToken
      parameters:
      - description: Token creation request
        name: request
        in: body
        required: true
        schema:
          $ref: '#/definitions/createTokenRequest'
      responses:
        "201":
          description: Token created successfully
          schema:
            $ref: '#/definitions/tokenCreateResponse'
        "400":
          description: Bad request error
          schema:
            $ref: '#/definitions/apiError'
          examples:
            application/json:
              message: Bad request
        "409":
          description: Token already exists
          schema:
            $ref: '#/definitions/apiError'
        "422":
          description: Chain not found or validation failed
          schema:
            $ref: '#/definitions/apiError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apiError'
          examples:
            application/json:
              message: Internal server error
...
  /api/v1/meta/{token_id}:
    get:
      description: 获取NFT的元数据信息，支持OpenSea标准格式
      tags:
      - nft
      summary: Get NFT metadata by token ID
      operationId: getNFTMetadata
      parameters:
      - type: string
        description: NFT Token ID
        name: token_id
        in: path
        required: true
      responses:
        "200":
          description: NFT metadata retrieved successfully
          schema:
            $ref: '#/definitions/nFTStandardMetadata'
        "404":
          description: NFT not found
          schema:
            $ref: '#/definitions/apiError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apiError'
          examples:
            application/json:
              message: Internal server error
...
definitions:
...
  ChainResponse:
    type: object
    properties:
      account_manager_address:
        description: Account manager contract address
        type: string
        example: 0xabcdefabcdefabcdefabcdefabcdefabcdefabcd
      chain_id:
        description: Chain ID
        type: integer
        format: int64
        example: 11155111
      cpop_token_address:
        description: CPOP token contract address
        type: string
        example: 0x1234567890123456789012345678901234567890
      created_at:
        description: Chain creation timestamp
        type: string
        format: date-time
        example: "2024-12-21T00:00:00Z"
      entry_point_address:
        description: EntryPoint contract address
        type: string
        example: 0x5FF137D4b0FDCD49DcA30c7CF57E578a026d2789
      explorer_url:
        description: Block explorer URL
        type: string
        example: https://sepolia.etherscan.io
      is_enabled:
        description: Whether chain is enabled
        type: boolean
        example: true
      master_aggregator_address:
        description: Master aggregator contract address
        type: string
        example: 0x0987654321098765432109876543210987654321
      max_batch_size:
        description: Maximum batch size
        type: integer
        example: 40
      min_batch_size:
        description: Minimum batch size
        type: integer
        example: 10
      name:
        description: Chain name
        type: string
        example: Sepolia Testnet
      optimal_batch_size:
        description: Optimal batch size for processing
        type: integer
        example: 25
      rpc_url:
        description: RPC endpoint URL
        type: string
        example: https://eth-sepolia.g.alchemy.com/v2/xxx
      short_name:
        description: Chain short name
        type: string
        example: sepolia
...
  apiError:
    type: object
    required:
    - message
    properties:
      errors:
        type: array
        items:
          type: object
          properties:
            code:
              description: Error code
              type: string
            field:
              description: Field name
              type: string
            message:
              description: Error message
              type: string
      message:
        description: Error message
        type: string
        example: Bad request
...
  chainDeleteResponse:
    type: object
    required:
    - message
    properties:
      message:
        type: string
        example: Chain deleted successfully
  chainToggleStatusResponse:
    type: object
    required:
    - message
    properties:
      message:
        type: string
        example: Chain status updated successfully
  chainUpdateResponse:
    type: object
    required:
    - message
    properties:
      message:
        type: string
        example: Chain configuration updated successfully
...
  createTokenRequest:
    type: object
    required:
    - chain_id
    - symbol
    - name
    - decimals
    - token_type
    properties:
      batch_operations:
        description: Supported batch operations configuration
        type: object
        additionalProperties: true
        example:
          batchBurn: false
          batchMint: false
          batchTransfer: false
      chain_id:
        description: Chain ID
        type: integer
        format: int64
        example: 56
      contract_address:
        description: Token contract address (null for native tokens)
        type: string
        example: 0x1234567890123456789012345678901234567890
      decimals:
        description: Token decimals
        type: integer
        example: 6
      name:
        description: Token name
        type: string
        example: Tether USD
      supports_batch_operations:
        description: Whether token supports batch operations
        type: boolean
        example: false
      symbol:
        description: Token symbol
        type: string
        example: USDT
      token_type:
        description: Token type
        type: string
        enum:
        - native
        - erc20
        example: erc20
...
  nFTStandardMetadata:
    type: object
    properties:
      attributes:
        type: array
        items:
          type: object
          properties:
            rarity_percentage:
              description: Rarity percentage
              type: number
              format: float
              example: 2.5
            trait_type:
              description: Attribute type
              type: string
              example: Background
            value:
              description: Attribute value
              type: string
              example: Galaxy
      description:
        description: NFT description
        type: string
        example: Friendly OpenSea Creature that enjoys long swims in the ocean.
      external_url:
        description: External URL
        type: string
        example: https://openseacreatures.io/3
      image:
        description: NFT image URL
        type: string
        example: https://storage.googleapis.com/opensea-prod.appspot.com/puffs/3.png
      name:
        description: NFT name
        type: string
        example: Dave Starbelly
...
  tokenCreateResponse:
    type: object
    required:
    - message
    - token_id
    properties:
      message:
        type: string
        example: Token created successfully
      token_id:
        type: integer
        example: 123
...
  tokenResponse:
    type: object
    properties:
      batch_operations:
        description: Supported batch operations configuration
        type: object
        additionalProperties: true
        example:
          batchBurn: true
          batchMint: true
          batchTransfer: true
      chain_id:
        description: Chain ID
        type: integer
        format: int64
        example: 56
      contract_address:
        description: Token contract address (null for native tokens)
        type: string
        example: 0x1234567890123456789012345678901234567890
      created_at:
        description: Token creation timestamp
        type: string
        format: date-time
        example: "2024-12-21T00:00:00Z"
      decimals:
        description: Token decimals
        type: integer
        example: 18
      id:
        description: Token ID
        type: integer
        example: 1
      is_enabled:
        description: Whether token is enabled
        type: boolean
        example: true
      name:
        description: Token name
        type: string
        example: CPOP Token
      supports_batch_operations:
        description: Whether token supports batch operations
        type: boolean
        example: true
      symbol:
        description: Token symbol
        type: string
        example: CPOP
      token_type:
        description: Token type
        type: string
        enum:
        - native
        - erc20
        example: erc20
  tokensResponse:
    type: array
    items:
      $ref: '#/definitions/tokenResponse'
...
  updateChainRequest:
    type: object
    properties:
      account_manager_address:
        description: Account manager contract address
        type: string
        example: 0xabcdefabcdefabcdefabcdefabcdefabcdefabcd
      cpop_token_address:
        description: CPOP token contract address
        type: string
        example: 0x1234567890123456789012345678901234567890
      entry_point_address:
        description: EntryPoint contract address
        type: string
        example: 0x5FF137D4b0FDCD49DcA30c7CF57E578a026d2789
      explorer_url:
        description: Block explorer URL
        type: string
        example: https://updated-explorer.example.com
      is_enabled:
        description: Whether chain is enabled
        type: boolean
        example: false
      master_aggregator_address:
        description: Master aggregator contract address
        type: string
        example: 0x0987654321098765432109876543210987654321
      max_batch_size:
        description: Maximum batch size
        type: integer
        example: 50
      min_batch_size:
        description: Minimum batch size
        type: integer
        example: 15
      name:
        description: Chain name
        type: string
        example: Updated Chain Name
      optimal_batch_size:
        description: Optimal batch size for processing
        type: integer
        example: 30
      rpc_url:
        description: RPC endpoint URL
        type: string
        example: https://updated-rpc.example.com
      short_name:
        description: Chain short name
        type: string
        example: updated
...
securityDefinitions:
  Bearer:
    description: |-
      Access token for application access, **must** include "Bearer " prefix.
      Example: `Bearer b4a94a42-3ea2-4af3-9699-8bcbfee6e6d2`
    type: apiKey
    name: Authorization
    in: header
    x-keyPrefix: 'Bearer '
  Management:
    description: Management secret, used for monitoring and infrastructure related
      calls
    type: apiKey
    name: mgmt-secret
    in: query